+++ src/core/server/effects/heal.ts
@@ -12,6 +12,7 @@ function handleItemEvent(player: alt.Player, item: Item, slot: number, tab: numb
     }
 
     playerFuncs.safe.addHealth(player, item.data.amount, false);
+    playerFuncs.inventory.notify(player, `+${item.data.amount} Health`);
 
     if (item.data.sound) {
         playerFuncs.emit.sound3D(player, item.data.sound, player);